eval/requirements.txt
# Requirements for MCP evaluator
pyyaml>=6.0
mcp-use>=0.1.0
anthropic>=0.39.0

# Optional: LLM providers (choose one)
# langchain-openai>=0.1.0  # For OpenAI
# langchain-anthropic>=0.1.0  # For Anthropic